Not grade

+++ srcs/cmd_readline_changes.c
@@ -6,7 +6,7 @@
 /*   By: sschmele <sschmele@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/01 13:38:26 by sschmele          #+#    #+#             */
-/*   Updated: 2019/09/23 17:42:07 by sschmele         ###   ########.fr       */
+/*   Updated: 2019/09/24 13:46:54 by sschmele         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -88,8 +88,8 @@ char		*help_str_change(char *cmd, char *swap, int point, char add)
 
 char		*str_add_symbol(char *cmd, char add, size_t *all)
 {
-	int		fact;
-	int		point;
+	size_t	fact;
+	size_t	point;
 	char	*swap;
 
 	fact = all[2] - PROMPT;
@@ -136,8 +136,8 @@ char		*str_add_symbol(char *cmd, char add, size_t *all)
 
 char		*str_del_symbol(char *cmd, size_t *all)
 {
-	int		fact;
-	int		point;
+	size_t	fact;
+	size_t	point;
 	char	*swap;
 
 	fact = all[2] - PROMPT;
